+++ src/engine/strat_engine/thinpool/thinpool.rs
@@ -1341,7 +1341,7 @@ mod tests {
     #[test]
     fn loop_test_full_pool() {
         loopbacked::test_with_spec(
-            &loopbacked::DeviceLimits::Exactly(2, Some(Bytes(IEC::Gi).sectors())),
+            &loopbacked::DeviceLimits::Exactly(2, Some(Bytes(u128::from(IEC::Gi)).sectors())),
             test_full_pool,
         );
     }
@@ -1351,8 +1351,8 @@ mod tests {
         real::test_with_spec(
             &real::DeviceLimits::Exactly(
                 2,
-                Some(Bytes(IEC::Gi).sectors()),
-                Some(Bytes(IEC::Gi * 4).sectors()),
+                Some(Bytes(u128::from(IEC::Gi)).sectors()),
+                Some(Bytes(u128::from(IEC::Gi) * 4).sectors()),
             ),
             test_full_pool,
         );
@@ -1641,7 +1641,7 @@ mod tests {
 
         // Create a filesystem as small as possible.  Allocate 1 MiB bigger than
         // the low water mark.
-        let fs_size = FILESYSTEM_LOWATER + Bytes(IEC::Mi).sectors();
+        let fs_size = FILESYSTEM_LOWATER + Bytes(u128::from(IEC::Mi)).sectors();
 
         let fs_name = "stratis_test_filesystem";
         let fs_uuid = pool
@@ -1668,7 +1668,7 @@ mod tests {
         }
         // Write 2 MiB of data. The filesystem's free space is now 1 MiB
         // below FILESYSTEM_LOWATER.
-        let write_size = Bytes(IEC::Mi * 2).sectors();
+        let write_size = Bytes(u128::from(IEC::Mi) * 2).sectors();
         let buf = &[1u8; SECTOR_SIZE];
         for i in 0..*write_size {
             let file_path = tmp_dir.path().join(format!("stratis_test{}.txt", i));
